pyproject.toml
[project]
name = "fastf1-oracleconverged"
version = "0.1.0"
description = "Oracle Converged Database demo using Formula 1 data - showcases relational, JSON, time series, spatial, graph, and vector capabilities"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastf1>=3.7.0",
    "pandas>=2.3.3",
    "oracledb>=2.0.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "folium>=0.15.0",
    "plotly>=5.18.0",
    "networkx>=3.2.0",
    "matplotlib>=3.8.0",
    "scikit-learn>=1.4.0",
    "kaleido>=0.2.1",
]

[project.scripts]
f1-env-check = "src.env_check:main"
